.ipynb_checkpoints/requirements-checkpoint.txt
Flask
Werkzeug
pytesseract
Pillow
pdf2image
Flask_SQLAlchemy